refactor: フィルタロジックを別ファイルに分離（パーツ一覧ページ）
state-serializer.ts が肥大化していたため、フィルタ関連のロジックを
filters.ts として独立したファイルに分離しました。

変更内容:
- filters.ts: フィルタロジックを新規作成
  - Filter 型定義
  - splitFiltersBySlot 関数（スロット切替時の条件分割）
  - applyFilters 関数（パーツへのフィルタ適用）
  - isValidFilterProperty 関数（プロパティ検証）
- filters.spec.ts: フィルタのテストを新規作成（13 tests）
- state-serializer.ts: フィルタロジックを削除し、filters.ts から再エクスポート
- state-serializer.spec.ts: フィルタのテストを削除（18 tests に整理）
- PartsListView.svelte: インポートを更新

テスト結果:
- filters.spec.ts: 13 tests passed
- state-serializer.spec.ts: 18 tests passed
- PartsListView.spec.ts: 15 tests passed

この変更により、関心の分離が明確化され、コードの保守性が向上しました。
